Patch 'Path' if imported from pathlib or pathlib2
- see #440
